+++ CHANGELOG.md
@@ -150,6 +150,5 @@ Version 1.2.5 - bug fixes and improvements.
 
 🐛 Fixed Bugs :
 - [x] Fix auto-number function: In 1.2.4 Auto-number still may incorrectly pushes the second and third tag renaming case, Now it will always correctly update only the first occurrence of tag.
-
-- [ ] (**Critical**) Fix citation update issue [bug #53](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/53), now citations in the current file will be updated correctly after auto-numbering. 
-
+- [x] (**Critical**) Fix citation update issue [bug #53](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/53), now citations in the current file will be updated correctly after auto-numbering.  
+- [ ] Prefix settings check